added static_cast to int conversion

+++ cpp02/ex01/srcs/Fixed.cpp
@@ -6,7 +6,7 @@
 /*   By: nlegrand <nlegrand@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/29 14:26:42 by nlegrand          #+#    #+#             */
-/*   Updated: 2023/08/02 18:15:58 by nlegrand         ###   ########.fr       */
+/*   Updated: 2023/08/03 13:34:00 by nlegrand         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,7 +30,7 @@ Fixed::Fixed(int const val)
 Fixed::Fixed(float const val)
 {
 	std::cout << "Float constructor called" << std::endl;
-	this->_val = roundf(val * (1 << this->_frac));
+	this->_val = static_cast<int>(roundf(val * (1 << this->_frac)));
 }
 
 // Destructors
